Put the grid back at the component level.
